moved location of filename_amender to helpers
