+++ src/estimation/estimation_main.py
@@ -5,7 +5,7 @@ import pandas as pd
 
 from src.estimation import apply_weights as appweights
 from src.estimation import calculate_weights as weights
-from src.utils.local_file_mods import filename_amender
+from src.utils.helpers import filename_amender
 
 EstMainLogger = logging.getLogger(__name__)
