+++ src/outputs/intram_by_civil_defence.py
@@ -2,7 +2,7 @@
 import logging
 import pandas as pd
 import numpy as np
-from src.utils.local_file_mods import filename_amender
+from src.utils.helpers import filename_amender
 from typing import Callable, Dict, Any
 
 OutputMainLogger = logging.getLogger(__name__)
