+++ src/outputs/intram_by_pg.py
@@ -1,7 +1,7 @@
 """The main file for the BERD Intram by PG output."""
 import logging
 import pandas as pd
-from src.utils.local_file_mods import filename_amender
+from src.utils.helpers import filename_amender
 from typing import Callable, Dict, Any
 
 OutputMainLogger = logging.getLogger(__name__)
