+++ src/outputs/intram_totals.py
@@ -4,7 +4,7 @@ import logging
 from typing import Callable, Dict, Any
 
 import pandas as pd
-from src.utils.local_file_mods import filename_amender
+from src.utils.helpers import filename_amender
 
 OutputMainLogger = logging.getLogger(__name__)
