+++ src/outputs/short_form.py
@@ -7,7 +7,7 @@ import src.outputs.map_output_cols as map_o
 from src.staging.validation import load_schema
 from src.imputation.imputation_helpers import fill_sf_zeros
 from src.outputs.outputs_helpers import create_output_df
-from src.utils.local_file_mods import filename_amender
+from src.utils.helpers import filename_amender
 
 
 OutputMainLogger = logging.getLogger(__name__)
