+++ src/utils/helpers.py
@@ -226,3 +226,19 @@ def tree_to_list(tree: dict, path_list: list = [], prefix: str = "") -> list:
         return path_list
     else:
         raise TypeError(f"Input must be a dictionary, but {type(tree)} is given")
+
+
+def filename_amender(filename, config):
+    """Append the survey type to filename if survey is PNG."""
+
+    survey_year = config["survey"]["survey_year"]
+    survey_type = config["survey"]["survey_type"]
+    run_id = config["filename_items"]["run_id"]
+    tdate = config["filename_items"]["tdate"]
+
+    if survey_type == "PNP":
+        filename = f"{survey_type}_{survey_year}_{filename}_{tdate}_v{run_id}.csv"
+    else:
+        filename = f"{survey_year}_{filename}_{tdate}_v{run_id}.csv"
+
+    return filename
